backend/requirements.txt
flask
flask-cors
firebase-admin
google-cloud-firestore
google-generativeai
google-auth
gunicorn
